noa-workshop/pyproject.toml
[project]
name = "noa-workshop"
version = "0.1.0"
description = "Telecom NOA-inspired multi-agent workshop using Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Microsoft Agent Framework — core SDK + Foundry provider
    "agent-framework>=0.5.0",
    # Foundry hosted-agent runtime (used for deployment in M5/extension)
    "agent-framework-foundry-hosting>=0.5.0",
    # Foundry Agent Service control-plane SDK (for FoundryAgent connection)
    "azure-ai-projects>=2.1.0",
    # Default credential chain (az login / managed identity)
    "azure-identity>=1.17.0",
    # Env loading
    "python-dotenv>=1.0.0",
    # OpenTelemetry console exporter for tracing demo (extension material)
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "agent-framework-devui>=1.0.0b260429",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/noa_workshop"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
